rename bedrock_channel_erosion to bedrock_channel
