+++ CMakeLists.txt
@@ -48,7 +48,7 @@ set(FASTSCAPELIB_HEADERS
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/consts.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/sinks.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/flow_routing.hpp
-  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/bedrock_channel_erosion.hpp
+  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/bedrock_channel.hpp
 )
 
 set(FASTSCAPELIB_TARGET fastscapelib)
