+++ include/fastscapelib/fastscapelib.hpp
@@ -6,5 +6,6 @@
  */
 #pragma once
 
+#include "fastscapelib/bedrock_channel.hpp"
 #include "fastscapelib/flow_routing.hpp"
 #include "fastscapelib/sinks.hpp"
